Refine syntax and start interpreting
Basic idea now is that we have a MetaInterpreter that has multiple
interpreters that run the programs for alternative executions.

A graph of actually-executed statements between then is called the
ExecutionLog and can be queried for a log between any two points like
log.between(Current, Past), or a tree [or reverse-tree] of all possible
logs to a single points with log.to(Current) or log.from(Current).

So the interpreter runs the actual program and serializes what it did in
a stream of statements. Then you can run that stream against another
interpreter and see if it works. I don't know yet how I will do
pause-and-continue in general. I'd need some way to specify where
exactly the program stopped. I guess some kind of stack trace.

Lots of things aren't implemented yet, like type checking, structs, and various
other things.