Start type checking
Nothing works so far, I'm just starting.

The general idea is that we have a checked variant of each
of the unchecked types in parser, e.g. `CheckedExpression`.

They are structured the same, and in the end hold `CheckedValue`
or `CheckedAccess`, and those hold the types they have. All
expressions will hold a type they have.

Statements won't hold a type, but will still have `Checked...` variants,
which simple indicate that they have been type checked.